requirements.txt
python-dateutil==2.8.2
pip==22.0.4
discord.py==1.7.3
python-dotenv==0.20.0